+++ kiapi/core/relay/__init__.py
@@ -1,3 +1,6 @@
+from ._exceptions.relay_request_error import RelayRequestError
+from ._helpers.build_relay_response import build_relay_response
+from ._helpers.new_relay_session_id import new_relay_session_id
 from ._instances.relay_registry import relay_registry
 from ._schemas.relay_file_body import RelayFileBody
 from ._schemas.relay_json_body import RelayJsonBody
@@ -25,10 +28,13 @@ __all__ = [
     "RelayMultipartFile",
     "RelayName",
     "RelayRequest",
+    "RelayRequestError",
     "RelayResponse",
     "RelayRunner",
     "RelaySettings",
     "RelaySpecifier",
+    "build_relay_response",
+    "new_relay_session_id",
     "relay_registry",
     "settings_manager",
 ]